client side of categories is completed as well

+++ client/client-admin/src/app/dashboard/categories/manage-categories/manage-categories.component.ts
@@ -1,5 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { CategoryService } from 'src/shared/services/category.service';
+import { NgxUiLoaderService } from 'ngx-ui-loader';
+import { BehaviorSubject } from 'rxjs';
+import { SnotifyService } from 'ng-snotify';
 
 @Component({
   selector: 'app-manage-categories',
@@ -8,9 +11,14 @@ import { CategoryService } from 'src/shared/services/category.service';
 })
 export class ManageCategoriesComponent implements OnInit {
 
-  constructor(private categoryService: CategoryService) { }
+  constructor(private categoryService: CategoryService, private ngxService: NgxUiLoaderService, private snotifyService: SnotifyService) { }
 
   categoriesList: any;
+  categoryQuery = "";
+
+  listLength = -1;
+
+  isLoadingQuery$ = new BehaviorSubject(false);
 
   async ngOnInit() {
     await this.getAllCategories();
@@ -18,37 +26,95 @@ export class ManageCategoriesComponent implements OnInit {
 
   async getAllCategories() {
     try {
+      this.ngxService.startBackground();
       return new Promise((resolve, reject) => {
         this.categoryService.getAllCategories()
           .subscribe((res) => {
             console.log(res);
             this.categoriesList = res['data'];
+            this.listLength = res['data'].length;
+            this.ngxService.stopBackground();
             resolve();
           }, (err) => {
             console.log('Categories not fetched', err);
+            this.ngxService.stopBackground();
             reject(err);
           })
       })
     } catch (err) {
       console.log('Internal Server Error', err);
+      this.ngxService.stopBackground();
+      this.snotifyService.error("Some internal server occured, kindly check after some time...");
     }
   }
 
   async deleteCategory(categoryData, index){
     try{
-      return new Promise((resolve, reject)=>{
-        this.categoryService.deleteCategory(categoryData)
-        .subscribe((res)=>{
-          this.categoriesList.splice(index, 1);
-          console.log('Category Deleted', res);
-          resolve();
-        }, (err) => {
-          console.log('Category not deleted', err);
-          reject(err);
-        })
+      this.snotifyService.confirm("Are you sure, you want to remoce this category?", {
+        timeout: 5000,
+        showProgressBar: true,
+        closeOnClick: true,
+        pauseOnHover: true,
+        buttons: [
+          {text: 'Yes', action: () => {
+            this.ngxService.startBackground();
+            return new Promise((resolve, reject)=>{
+              this.categoryService.deleteCategory(categoryData)
+              .subscribe((res)=>{
+                this.categoriesList.splice(index, 1);
+                console.log('Category Deleted', res);
+                this.listLength = this.categoriesList.length;
+                this.ngxService.stopBackground();
+                this.snotifyService.info("Category removed!");
+                resolve();
+              }, (err) => {
+                console.log('Category not deleted', err);
+                this.ngxService.stopBackground();
+                this.snotifyService.error("Failed to remove category, please try again!");
+                reject(err);
+              })
+            })
+          }, bold: false},
+          {text: 'No', action: (toast) => { this.snotifyService.remove(toast.id); }, bold: true},
+        ]
       })
     } catch(err){
       console.log('Internal Server Error', err);
+      this.ngxService.stopBackground();
+      this.snotifyService.error("Some internal server occured, kindly check after some time...");
+    }
+  }
+
+  async searchQuery(categoryName){
+    try{
+      if(this.listLength > 0){
+        this.ngxService.startBackground();
+        return new Promise((resolve, reject) => {
+          this.categoryService.searchCategory(categoryName)
+          .subscribe((res)=>{
+            console.log('Categories Found', res);
+            this.categoriesList = res['data'];
+  
+            if(res['data'].length == 0)
+              this.isLoadingQuery$.next(true);
+            else
+              this.isLoadingQuery$.next(false);
+  
+            this.ngxService.stopBackground();
+            resolve();
+          }, (err) =>{
+            console.log('Categories not found', err);
+            this.ngxService.stopBackground();
+            this.isLoadingQuery$.next(false);
+            reject(err);
+          })
+        })
+      }
+    } catch(err){
+      console.log('Internal Server Error', err);
+      this.isLoadingQuery$.next(false);
+      this.ngxService.stopBackground();
+      this.snotifyService.error("Some internal server occured, kindly check after some time...");
     }
   }
